+++ ios/IJKMediaPlayer/IJKMediaPlayer/IJKFFMoviePlayerController.m
@@ -34,7 +34,7 @@
 #include "ijkplayer/version.h"
 #include "ijkplayer/ijkavformat/ijkavformat.h"
 
-static const char *kIJKFFRequiredFFmpegVersion = "ff2.8--ijk0.4.4.1--dev0.3.3--rc7";
+static const char *kIJKFFRequiredFFmpegVersion = "ff3.0--ijk0.4.5.1--dev0.4.5--rc10";
 
 @interface IJKFFMoviePlayerController()